separate model, controller, db connection
