+++ .env
@@ -1,0 +1,7 @@
+API_KEY=your_api_key
+
+DATABASE_HOST=localhost
+DATABASE_PORT=5432
+DATABASE_USER=postgres
+DATABASE_PASSWORD=root
+DATABASE_NAME=money-tracker-db
