+++ src/middleware/jwt.go
@@ -1,0 +1,42 @@
+package middleware
+
+import (
+	"net/http"
+
+	"github.com/dgrijalva/jwt-go"
+	"github.com/gin-gonic/gin"
+)
+
+// Claims struct to handle JWT claims
+type Claims struct {
+	Email string `json:"email"`
+	jwt.StandardClaims
+}
+
+// JWTMiddleware validates JWT tokens from cookies
+func JWTMiddleware(jwtKey []byte) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cookie, err := c.Cookie("token")
+		if err != nil {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
+			c.Abort()
+			return
+		}
+
+		claims := &Claims{}
+
+		token, err := jwt.ParseWithClaims(cookie, claims, func(token *jwt.Token) (interface{}, error) {
+			return jwtKey, nil
+		})
+		if err != nil || !token.Valid {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
+			c.Abort()
+			return
+		}
+
+		// Set claims into gin context for further use
+		c.Set("claims", claims)
+
+		c.Next()
+	}
+}
