+++ src/model/user.go
@@ -1,0 +1,8 @@
+package models
+
+type User struct {
+	ID       int    `db:"id" json:"id"`
+	Username string `db:"username" json:"username"`
+	Password string `db:"password" json:"password"`
+	Email    string `db:"email" json:"email"`
+}
